Beautify code

+++ common.cpp
@@ -4,10 +4,10 @@
 #include "common.h"
 #include "utils.h"
 
-int recv_all(int sockfd, void *buffer, size_t len) {
+int recv_all(int sockfd, void *buffer, int len) {
 
-	size_t bytes_received = 0;
-	size_t bytes_remaining = len;
+	int bytes_received = 0;
+	int bytes_remaining = len;
 	char *buff = (char *)buffer;
 
   	while (bytes_remaining > 0) {
@@ -24,9 +24,9 @@ int recv_all(int sockfd, void *buffer, size_t len) {
 	return bytes_received;
 }
 
-int send_all(int sockfd, void *buffer, size_t len) {
-	size_t bytes_sent = 0;
-	size_t bytes_remaining = len;
+int send_all(int sockfd, void *buffer, int len) {
+	int bytes_sent = 0;
+	int bytes_remaining = len;
 	char *buff = (char *)buffer;
 
 	while (bytes_remaining > 0) {
